requirements.txt
pytorch==1.0.0
torchvision == 0.2.1
Pillow==7.2.0
tensorboardX==2.0
numpy==1.17.2
tqdm